app/src/pyproject.toml
[tool.poetry]
name = "skyengtesttask"
version = "0.1.0"
description = "Python file checker service"
authors = ["Alexander Zhabinskiy <alexzhabinskiy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
psycopg = {extras = ["binary"], version = "^3.1.10"}
gunicorn = "^21.2.0"
django-allauth = "^0.54.0"
crispy-tailwind = "^0.5.0"
celery = "^5.3.1"
flake8 = "^6.1.0"
django-celery-beat = "^2.5.0"
pygments = "^2.16.1"
bleach = "^6.0.0"
whitenoise = {extras = ["brotli"], version = "^6.5.0"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
django-debug-toolbar = "^4.2.0"
pylint = "^2.17.5"
pylint-django = "^2.5.3"
pytest-django = "^4.5.2"
pytest-celery = "^0.0.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Configuration for various linters, formatters, checkers etc.
[tool.black]
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
filter_files = true

[tool.pylint.messages_control]
max-line-length = 88
score=false
reports=false
disable = [
    "missing-docstring",
    "too-few-public-methods",
    "too-many-ancestors"
]

[tool.pylint.MASTER]
ignore-paths = [
    "^app/src/.*/migrations/.*$",
    "^app/src/manage.py$",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
FAIL_INVALID_TEMPLATE_VARS = true
